+++ src/fc/Guest.java
@@ -1,12 +1,13 @@
 package fc;
 
+import java.math.BigInteger;
 import java.util.HashSet;
 
 public class Guest extends Personne {
 	private LocationBR locBr;
 	private HashSet<LocationQR> locQr;
 	
-	public Guest(int cb) {
+	public Guest(BigInteger cb) {
 		super(0,cb);
 		this.locQr = new HashSet<>();
 	}
